% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204159.R
\name{lung}
\alias{lung}
\title{A illustration dataset}
\description{
A dataset which contained information of patients who have lung cancer .This dataset is used to illustrate the how to construct a XGBoost prognostic model.The last two column is the endpoint status and survival months of patient,Status 1 represents the patient is censored at endpoint and status 2 represents the patient is dead at endpoint.The rest of colmuns are the covariate.
}
\examples{
\dontrun{
data(lung)
attach(lung)
head(lung)
}
}
