% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204159.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{A function used to visualize confusion matrix}
\usage{
plot_confusion_matrix(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A confusion matrix}
}
\description{
A function used to visualize confusion matrix
}
\examples{
\dontrun{
data(lung)
attach(lung)
res<-XGB_md(6,lung)
plot_confusion_matrix(res$confusion_matrix)
}
}
