% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204159.R
\name{plot_roc}
\alias{plot_roc}
\title{A function used to plot roc curve}
\usage{
plot_roc(roc_value)
}
\arguments{
\item{roc_value}{a series of prediction probility}
}
\description{
A function used to plot roc curve
}
\examples{
\dontrun{
data(lung)
attach(lung)
res<-XGB_md(6,lung)
plot_roc(res$prob_train)
}
}
