\name{SA23204159-package}
\alias{SA23204159-package}
\alias{SA23204159}
\docType{package}
\title{
  Final project for SA23204159
}
\description{
  Several R and Rcpp function are included in this package for illustration.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Zhijie Gong

Maintainer: Zhijie Gong <1218469181@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
