% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{studentized_bootstrap_ci_C}
\alias{studentized_bootstrap_ci_C}
\title{construct studentized bootstrap CI}
\usage{
studentized_bootstrap_ci_C(X, B = 500L)
}
\arguments{
\item{X}{the variable that you want to construct CI,must be a vector}

\item{B}{the number of times that replicate}
}
\value{
lower and upper bound of CI
}
\description{
construct studentized bootstrap CI
}
\examples{
\dontrun{
data(lung)
attach(lung)
##construct CI for patients' survival month,use lung data
data<-lung$存活时间
studentized_bootstrap_ci(data)
}
}
