% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204159.R
\name{XGB_md}
\alias{XGB_md}
\title{Train a predictive model use XGBoost algorithm}
\usage{
XGB_md(n_month, data, seed_number = 12345)
}
\arguments{
\item{n_month}{the number of month that we want to predict the survival of patient}

\item{data}{the dataset that we use to construct XGBoost prognostic model.The data should in the form that last columns is the endpoint of status and}

\item{seed_number}{the number of seed that we want to set}
}
\value{
a list contains prob_train, prob_test, confusionmatrix, md. Prob_train:predicted probility of dead at endpoint on train data,prob_test:predicted probility of dead at endpoint on test data,confusion_matrix:confusion matrix on test data,md:predictive model
}
\description{
Train a predictive model use XGBoost algorithm
}
\examples{
\dontrun{
data(lung)
attach(lung)
XGB_md(6,lung)
}
}
